rename method and move to utils
trigger ci
